patch Lua, use portable bytecode

+++ lib/cocos2d-x/scripting/lua/lua/src/ldump.c
@@ -62,12 +62,14 @@ static void DumpString(const TString* s, DumpState* D)
 {
  if (s==NULL)
  {
-  size_t size=0;
+  // size_t size=0;
+  unsigned int size=0;
   DumpVar(size,D);
  }
  else
  {
-  size_t size=s->tsv.len+1;		/* include trailing '\0' */
+  // size_t size=s->tsv.len+1;		/* include trailing '\0' */
+  unsigned int size=(unsigned int)s->tsv.len+1;
   DumpVar(size,D);
   DumpBlock(getstr(s),size*sizeof(char),D);
  }
